Make on-demand base capacity configurable
Currently, we set the on-demand base capacity to the same value as the
minimum-workers-per-az-count.  But in sandbox I want to set the base
capacity to 0 (so we can have 100% spot instances).

This makes the on-demand base capacity separately configurable, with a
default value of 1 (which matches the current value everywhere).

:warning: I think this will respin all the worker nodes because it
edits the launch template in the cloudformation template.