Don't leave EBS volumes lying around when draining
Currently, when we drain a cluster, we leave a bunch of EBS volumes
lying around.  This is because we don't delete the persistentvolumes
correctly.  This is expensive, because each time we do this we leak
~600GB worth of EBS space (of which 500GB is prometheus).  EBS costs
$0.116 per GB-month in eu-west-2, so 600GB works out around $70/month
worth of unused EBS leaked each time we drain a cluster.

In order to do delete EBS properly, we need to delete any pod that has
a bound persistentvolume first.

Therefore, this commit:

 - deletes all deployments and statefulsets (in order to delete the
   pods)
 - deletes all persistentvolumes and persistentvolumeclaims

This should hopefully delete the underlying EBS volumes.