Delete concourse pipelines when we drain clusters
there is a bug in the concourse-operator that will cause Pipeline
resources to become un-deletable if their parent Team resource is
deleted first. So you must delete Pipelines _before_ deleteing Teams.

this bug manifests it's during draining as we delete all the CRDs (wihch
cascades to deleting the CRs), and can cause Teams to be deleted before
Pipelines and getting stuck.

as a workaround to this bug we delete all the Pipelines early, avoiding
the ordering issue when the CRDs get deleted.