ignore errors deleteing concourse Team resources
concourse's team model currently requires at least one "admin" team
(usually called "main") because authenication is tied to a team.

the concourse-operator requires that the main exists as the operator
credentials are for the "main" team.

we have had issues where the deletion of the main team causes other
operator faliures and Team resources getting stuck in deleteing states.

this workaround is a bit lazy (only logs team destroy errors), but:

1. in the future there are plans to decouple concourse auth from teams,
   which would make this problem much easier to solve / moot
2. teams are rarely destroyed (except with on-demand clusters) so it's
   not worth the engineering effort right now to resolve it nicely