support spot termination events in lifecycle hook
we use spot instances in some clusters, but spot instances do not obey
the rules of lifecycle hooks.

to ensure more graceful handling of spot termination we extend the
lifecycle hook to perform the same drain functionality currently
executed for ASG termination events for the spot termination warning
events.

In additional to wiring up the events, we also slightly modify the drain
logic to not error if node cannot be found (this is because we cannot
scope the spot termination warning events to the correct asg easily, so
we need to gracefully handle the case where the hook is called for a
node in a different cluster - not ideal, but not so bad).