Create cloudwatch logs VPC endpoint
It seems this should be enough to get logs shipped via the endpoint rather than
the existing mechanism. This changes DNS resolutions within the cluster to have
this:
(after apt update and apt install dnsutils)
root@gsp-fluentd-cloudwatch-8nzph:/home/fluent# dig +short logs.eu-west-2.amazonaws.com
10.101.24.69
10.101.133.238
10.101.93.108

Whereas externally you get a non-RFC1918 IP like 52.94.48.143.

And the cluster seems to trust the cert served etc. Logs appear to be getting
shipped in my cluster.
TODO: Do we want to hide this behind a flag to enable it on a per-cluster basis?