enable 14days of vpc flow logs
we are seeing some higher than expected costs due to traffic through our
NAT gateways.

This enables collection of the VPC network traffic flow logs with 14 day
retention so that we can identify which service(s), if any, are
disproportionally responsible for the network transfer costs through the
gateways.

The aim is to create VPC endpoints for any services that may be
responsible to improve routing.

This change is considered temporary while we investigate. If your are
reading this commit in the future wondering if you can turn off the VPC
flow logs... please do!