do not update role annotation unless present
there is an edge case where an error in the cloudformation status update
can mark the status as "ready" but have failed to update the info fields
that the service account controller relies on.

It's not obvious if this should be fixed in the status update code, as
the intention of that code is to reflect as much of the cloudformation
state in the resource as possible within each reconciliation loop.

It _is_ clear however that the service account annotation should always
be expected to be present, so if it is blank, then that is an error and
it should fail to reconcile