Ignore changes to current running worker counts
We have seen issues where, during a terraform apply, not all the nodes
will roll. This appeared to be tied to availability zone - all nodes in
an AZ worker group would roll but none of the others would. We're
working on the hypothesis that the trigger is the change in the
"desired" count that is caused by the cluster autoscaler. By ignoring
this value we hope to avoid unnecessary node rolls.