[untested] work around concourse volume problem in deployer pipeline
I think we've hit a concourse bug when we upgraded from concourse
6.0.0 to 6.3.0.  I've filed concourse/concourse#5799 upstream to
capture this.  In the meantime, I think we can work around this by
forcing each task to use a different volume.

I'm hoping that, by using the `path:` parameter on `output`
configuration blocks, I can leave the actual task steps untouched; I'm
outputting the same directory, but to a different volume.

I don't know how this interacts with `put` steps (in particular, when
we `put: cluster-state`).

Let's merge and see if it fixes things?  It can't break them more than
they are already...