canary: trigger deploy from timer
we are seeing the "build" job run more often than the "deploy" job in
the smoke test pipeline.

Sometimes the "deploy" job waits around for some time (~10mins) before triggering,
resulting in the CanaryOverdue alert firing.

It's not 100% clear why this is happening, and an investigation could be
yak shavey at this point, but as a quick experiement this changes the
job to be triggered on the timer resource rather than the ecr repository
in case the issue is with the check loop of the ecr resource and
increases the interval a little in case the issue is with some kind of
locking while the deploy job is "putting" to the ECR resource (total
guess work).