Lock the engine version of RDS aurora-postgres
Without setting the version RDS defaults to "latest", which isn't in line
with our parameter group family, which does lock in the version. So,
explicitly lock the engine version to keep it consistent with the parameter
group.

An alternative would be to expose some sort of parameter through the yaml
and align accordingly, but given the EOL mode of GSP this is the simplest
change that will work.