concourse-op: stop using id-token as auth-token
it appears we were pulling the id-token from the JWT token response and
using it as the bearer token for the client. I'm not sure why we were
doing this, but it appears to have been introduced shortly after the
concourse authentication changes in v6.2. I suspect whatever the
original problem was has not been fixed, and so we ended up using the
wrong token after the upgrade to 6.6.

using the standard AuthToken response resolves our current issues with
the password authenticate for the operator not working against v6.6

this change likely makes the operator no-longer compatible with versions
prior to 6.6