concourse-op: fix config unmarshaling for Pipeline
Recent changes in the concourse atc.Config type have caused an
incompatibility between how the k8s apimachinry wants to unmashal the
resource spec into the type and how concourse expects atc.Config to be
unmarshaled.

Essentially previously a simple json.Unmarshal would have populated an
atc.Config correctly, whereas now the dedicated atc.UnmarshalConfig is
the only way to get a reliable decode.

To solve this, I have wrapped the Pipeline.Spec.Config in a type of our
own that implements the json.Unmarshaler interface to catch the decoding
and populate the Config as expected ... this is a bit ugly, but it was
simpler than attempting to catch the decode on the Spec and have to pick
the fields apart.

I may raise a PR upstream to implement Unmarshaler on atc.Config
directly as this seems like the sensible thing to do.

While attempting to reproduce the issue locally I re-jigged the tests a
little so that they work at the Resource level, making it clearer whats
going on.

And to avoid confusion with the different yaml encoding/decoding going
on I have dropped the use of yaml.v2 lib and used the k8s.io/yaml lib
(same one used in the apimachinery)