Provide DockerHub credentials to tenants and various system components
To avoid rate limiting problems.

These will be different credentials to those used in the release pipeline -
these will be shared with tenants and therefore will not be given any special
push permissions, they will be used for pulling things like Istio containers,
and other public images from DockerHub.

Provide it in the Docker config format for Kubernetes, and a secret with
username and password for use in Concourse.